backend/requirements.txt

Flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.0.1
python-docx==0.8.11
gunicorn==21.2.0